Removed unused code and moved “docker_client” code to image package.
